pyproject.toml
[project]
name = "skyweb"
version = "0.1.0"
description = "VFR flight preparation automation tool"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.6,<3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.3",
    "mypy>=1.9",
]
api = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "python-multipart>=0.0.9",
    "httpx>=0.27",
]
gcp = [
    "google-cloud-firestore>=2.16",
    "google-cloud-storage>=2.16",
    "firebase-admin>=6.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
